Make parent class, fixed file-loader
