+++ ST/Kirsanov.Ivan/Project/public/index.html
@@ -536,7 +536,7 @@
                         <div>hot deal</div>
                         <div>for men</div>
                     </a>
-                    <a href="#"><img src="../src/assets/imgs/offer1.jpg" alt=""></a>
+                    <a href="#"><img src="https://raw.githubusercontent.com/kellolo/static/master/img/JS1_shop/offer1.jpg" alt=""></a>
                 </div>
 
                 <div class="offer mt-4">
@@ -544,7 +544,7 @@
                         <div>LUXIROUS&amp;trendy</div>
                         <div>ACCESORIES</div>
                     </a>
-                    <a href="#"><img src="../src/assets/imgs/offer2.jpg" alt=""></a>
+                    <a href="#"><img src="https://raw.githubusercontent.com/kellolo/static/master/img/JS1_shop/offer2.jpg" alt=""></a>
                 </div>
             </div>
 
@@ -554,7 +554,7 @@
                         <div>30% offer</div>
                         <div>women</div>
                     </a>
-                    <a class="one" href="#"><img src="../src/assets/imgs/offer3.jpg" alt=""></a>
+                    <a class="one" href="#"><img src="https://raw.githubusercontent.com/kellolo/static/master/img/JS1_shop/offer3.jpg" alt=""></a>
                 </div>
 
                 <div class="offer pt-4 pt-sm-0 order-1 order-md-2 mt-0 mt-md-4">
@@ -562,7 +562,7 @@
                         <div>new arrivals</div>
                         <div>FOR kids</div>
                     </a>
-                    <a class="one" href="#"><img src="../src/assets/imgs/offer4.jpg" alt=""></a>
+                    <a class="one" href="#"><img src="https://raw.githubusercontent.com/kellolo/static/master/img/JS1_shop/offer4.jpg" alt=""></a>
                 </div>
             </div>
         </div>
