+++ ST/Kirsanov.Ivan/Project/src/components/items.js
@@ -1,12 +1,16 @@
-export default class BasketItem {
+import Item from './listItem';
+
+export class CatalogItem extends Item {}
+
+export class BasketItem extends Item {
     constructor(item){
-        this.item = item;
+        super(item);
     }
     
     render() {
         return `
             <div class="d-flex headerCartWrapIn mb-1 p-2">
-                    <img src="${this.item.productImg}" alt="" width="85" height="100>
+                    <img src="${this.item.productImg}" alt="" width="85" height="100">
                     <div>
                         <div>${this.item.productName}</div>
                         <span>
@@ -19,13 +23,14 @@ export default class BasketItem {
                         <div class="headerCartWrapPrice">${this.item.amount} 
                             <span>x</span> $${this.item.productPrice}
                         </div>
-
+                    </div>
                 <button 
                     class="fas fa-times-circle" 
                     data-id="${this.item.productId}"
                     name="remove"
                 ></button>
             </div>
+            
         `
     }
 }
